% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BayesFactorExtras-package.R
\docType{package}
\name{BayesFactorExtras-package}
\alias{BayesFactorExtras}
\alias{BayesFactorExtras-package}
\title{The BayesFactorExtras package}
\description{
Extra functions for use with the BayesFactor package
}
\details{
This package contains extra functions that can be used to enhance the BayesFactor
package, such as plots and reports.

\tabular{ll}{ Package: \tab BayesFactorExtras\cr Type: \tab Package\cr Version: \tab
0.1.0\cr Date: \tab 2014-11-22\cr License: \tab GPL 2.0\cr LazyLoad: \tab
yes\cr }
}
\examples{
## See specific functions for examples.
}
\author{
Richard D. Morey

 Maintainer: Richard D. Morey <richarddmorey@gmail.com>
}
\seealso{
\code{\link[BayesFactor:BayesFactor-package]{BayesFactor}}
}
\keyword{htest}

