% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDownloadURI}
\alias{createDownloadURI}
\title{Download R data files via HTML}
\usage{
createDownloadURI(list, filename = stop("'filename' must be specified"),
  textHTML = "Click here to download the data.", fileext = "RData",
  envir = parent.frame())
}
\arguments{
\item{list}{R objects to put in the Rdata file}

\item{filename}{A name to give the downloaded file (without extension)}

\item{textHTML}{text to link in the web page}

\item{fileext}{an extension to give the file}

\item{envir}{the environment to search for the objects in \code{list}}
}
\value{
This function returns NULL invisibly.
}
\description{
Creates HTML to insert Rdata file into web page
}
\details{
Create HTML to insert Rdata file into web page. This is convenient for use with
\code{brew} or \code{knitr}.
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\references{
See \url{http://bayesfactor.blogspot.nl/2014/09/embedding-rdata-files-in-rmarkdown.html}
}
\keyword{misc}

