% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interactBF}
\alias{interactBF}
\title{View interactive BayesFactor object}
\usage{
interactBF(BFObj)
}
\arguments{
\item{BFObj}{BayesFactor object}
}
\value{
This function returns the html file name and path
}
\description{
Creates interactive HTML BayesFactor and views with default viewer
}
\details{
Creates interactive HTML BayesFactor and views with default viewer
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\keyword{misc}

