% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/seqBF.R
\name{seqBF}
\alias{seqBF}
\title{Compute a sequential Bayes factor object}
\usage{
seqBF(BFobject, min.n = 10, step = 1, var.id = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{BFobject}{A BayesFactor object}

\item{min.n}{Smallest number of participants for which the first BF should be computed. In a one-sample t test / paired  test, this is the overall sample size. In two-sample t tests and ANOVAs, this is the minimum size for each group. #}

\item{step}{Step size with which the number of participants is increased}

\item{var.id}{The column that defines which rows belong to one person/object of observation. This is relevant for repeated measurements, as we want to sample complete observations. If this parameter is left as NULL, the function tries to guess the relevant identifying variable.}

\item{verbose}{Print diagnostic information?}

\item{...}{User parameters which override the settings of the original \code{BFobject}}
}
\description{
Compute a sequential Bayes factor object
}
\details{
Compute a sequential Bayes factor object
}
\examples{
# Paired/ one-sample t test
ttest.p <- ttestBF(x = sleep$extra[sleep$group==1], y = sleep$extra[sleep$group==2], paired=TRUE)
ttest.p.seq <- seqBF(ttest.p, min.n=5)
plot(ttest.p.seq)

# AnovaBF
anova1 <- anovaBF(extra ~ group + ID, data = sleep, whichRandom = "ID", progress=FALSE)
anova1.seq <- seqBF(anova1, min.n=5)
plot(anova1.seq)

data(puzzles)
(anova2 <- anovaBF(RT ~ shape*color + ID, data = puzzles, whichRandom = "ID", progress=FALSE))
(anova2.seq <- seqBF(anova2, min.n=5))
plot(anova2.seq)
}
\author{
Felix Schönbrodt
}

