% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqBFplot.R
\name{seqBFplot}
\alias{seqBFplot}
\title{Plot a Bayes factor object}
\usage{
seqBFplot(n, bf, linetype = NA, linetype.label = "Alternative",
  xlab = "n", ylab = "Bayes Factor", main = "", ylim = NA,
  log.it = TRUE, forH1 = TRUE, fontsize = 3.2)
}
\arguments{
\item{n}{A vector of numbers for the x axis}

\item{bf}{A vector of Bayes factors (same length as x)}

\item{linetype}{If several lines should be drawn, this variable defines which data point belongs to which line}

\item{linetype.label}{The heading of the linetype legend}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{main}{Main title}

\item{ylim}{Limits of y axis, in raw BF units. If NA, plot is automatically adjusted to range of data.}

\item{log.it}{Should the Bayes factor in the \code{bf} parameter be logged?}

\item{forH1}{If \code{TRUE}, positive BFs mean evidence in favor of H1 ("H1 over H0" Bayes factor). This is the default in the Bayes factor package.}

\item{fontsize}{Controls the font size of the annotation}
}
\description{
Plot a sequence of Bayes factors
}
\details{
This function creates a lineplot of a sequence of log Bayes factors. As a default, the function expects raw Bayes factors (i.e., non-logged Bayes factors).
If you provide Bayes factors that are already logged (e.g., from the output of the \code{\link{ttest.tstat}} function), set \code{log.it} to \code{FALSE}.
This function is in particular useful for plotting the trajectory of a sequential Bayes factor test
}
\examples{
## Sleep data from t test example
data(sleep)

# Compute accumulating evidence from n1=5 participants to n2=10 participants
bf <- c()
for (i in 5:10) {
	bf0 <- ttestBF(
		x = sleep$extra[sleep$group==1][1:i],
		y = sleep$extra[sleep$group==2][1:i], paired=TRUE)
	bf <- c(bf, as.vector(bf0))
}

seqBFplot(5:10, bf)
seqBFplot(1:100, cumsum(rnorm(100, 2)))
}
\author{
Felix D. Schönbrodt (\email{felix@nicebread.de})
}

