% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactBF.R
\name{interactBF}
\alias{interactBF}
\title{View interactive BayesFactor object}
\usage{
interactBF(BFObj)
}
\arguments{
\item{BFObj}{BayesFactor object}
}
\value{
This function returns the html file name and path
}
\description{
Creates interactive HTML BayesFactor and views with default viewer
}
\details{
Creates interactive HTML BayesFactor and views with default viewer
}
\examples{
## Create a BayesFactor analysis
data(puzzles)
bf = anovaBF(RT ~ shape*color + ID, data = puzzles, whichRandom = "ID",
    progress=FALSE)
\dontrun{
## Open browser with interactive BayesFactor analysis
interactBF(bf)
}
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\keyword{misc}

