% Generated by roxygen2 (4.0.2): do not edit by hand
\name{knit_print.BFBayesFactor}
\alias{knit_print.BFBayesFactor}
\title{Print interactive HTML/JavaScript Bayes factors}
\usage{
\method{knit_print}{BFBayesFactor}(x, ...)
}
\arguments{
\item{x}{a BFBayesFactor object}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
This function code to be inserted into Rmarkdown documents.
}
\description{
knitr print method for BFBayesFactor objects
}
\details{
knitr print method for BFBayesFactor objects
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}

