% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onAttach.R
\name{options-BayesFactorExtras}
\alias{options-BayesFactorExtras}
\title{options() for package BayesFactorExtras}
\description{
Options that can be set for the BayesFactorExtras package
}
\details{
The BayesFactorExtras package has options that can be set to globally
change the behavior of the functions in the package. These options can be
changed using \code{\link[base]{options}}().

\describe{
\item{\code{BFEknitrDownload}}{If \code{TRUE}
(default), a link to download the Bayes factor object is
provided. This object contains the analysis and the data.}
}
}
\seealso{
\code{\link[base]{options}}
}

