% Generated by roxygen2 (4.0.2): do not edit by hand
\name{posterior.plot}
\alias{posterior.plot}
\title{Plot posterior from BayesFactor object}
\usage{
posterior.plot(object, ...)
}
\arguments{
\item{object}{a BayesFactor object}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
This function returns NULL
}
\description{
Plot the posterior (and possibly prior) from a BayesFactor object
}
\details{
Plot the posterior (and possibly prior) from a BayesFactor object
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\keyword{hplot}

