% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seqBFplot}
\alias{seqBFplot}
\title{Plot a Bayes factor object}
\usage{
seqBFplot(n, BF, xlab = "n", ylab = "log(BF)", main = "", log.it = TRUE,
  forH1 = TRUE)
}
\arguments{
\item{n}{A vector of numbers for the x axis}

\item{BF}{A vector of Bayes factors (same length as x)}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{main}{Main title}

\item{log.it}{Should the Bayes factor in the \code{BF} parameter be logged?}

\item{forH1}{If \code{TRUE}, positive BFs mean evidence in favor of H1 ("H1 over H0" Bayes factor). This is the default in the Bayes factor package.}
}
\description{
Plot a sequence of Bayes factors
}
\details{
This function creates a lineplot of a sequence of log Bayes factors. As a default, the function expects raw Bayes factors (i.e., non-logged Bayes factors).
If you provide Bayes factors that are already logged (e.g., from the output of the \code{\link{ttest.tstat}} function), set \code{log.it} to \code{FALSE}.
This function is in particular useful for plotting the trajectory of a sequential Bayes factor test
}
\examples{
## Sleep data from t test example
data(sleep)

# Compute accumulating evidence from n1=5 participants to n2=10 participants
BF <- c()
for (i in 5:10) {
	BF0 <- ttestBF(
		x = sleep$extra[sleep$group==1][1:i],
		y = sleep$extra[sleep$group==2][1:i], paired=TRUE)
	BF <- c(BF, as.vector(BF0))
}

seqBFplot(5:10, BF)
}
\author{
Felix D. Schönbrodt (\email{felix@nicebread.de})
}

